New full screen chat WIP + code refactoring.
