+++ assets/styles/style.scss
@@ -169,11 +169,3 @@ table.peertube-plugin-livechat-prosody-list-rooms td {
     min-height: 60px;
   }
 }
-
-.livechat-embed-fullpage {
-  converse-root {
-    display: block;
-    height: 90vh;
-    min-height: 400px;
-  }
-}
